Convert Java to use doubles instead of decimals.

+++ java/src/main/java/com/granularenergy/allocation/models/ConsumptionVolume.java
@@ -1,11 +1,10 @@
 package com.granularenergy.allocation.models;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ConsumptionVolume(
     String consumer,
     LocalDateTime periodStart,
     Duration duration,
-    BigDecimal quantityKwh
+    double quantityKwh
 ) {}
