+++ java/src/main/java/com/granularenergy/allocation/models/ProductionVolume.java
@@ -1,11 +1,10 @@
 package com.granularenergy.allocation.models;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProductionVolume(
     String producer,
     LocalDateTime periodStart,
     Duration duration,
-    BigDecimal quantityKwh
+    double quantityKwh
 ) {}
